Record tids in the thread's own PID namespace and print them during rr ps
