+++ src/TraceStream.cc
@@ -32,7 +32,7 @@ namespace rr {
 // MUST increment this version number.  Otherwise users' old traces
 // will become unreplayable and they won't know why.
 //
-#define TRACE_VERSION 68
+#define TRACE_VERSION 69
 
 struct SubstreamData {
   const char* name;
@@ -294,7 +294,7 @@ void TraceWriter::write_task_event(const TraceTaskEvent& event) {
   tasks << global_time << (char)event.type() << event.tid();
   switch (event.type()) {
     case TraceTaskEvent::CLONE:
-      tasks << event.parent_tid() << event.clone_flags();
+      tasks << event.parent_tid() << event.own_ns_tid() << event.clone_flags();
       break;
     case TraceTaskEvent::EXEC:
       tasks << event.file_name() << event.cmd_line() << event.fds_to_close();
@@ -317,7 +317,7 @@ TraceTaskEvent TraceReader::read_task_event() {
   r.type_ = (TraceTaskEvent::Type)type;
   switch (r.type()) {
     case TraceTaskEvent::CLONE:
-      tasks >> r.parent_tid_ >> r.clone_flags_;
+      tasks >> r.parent_tid_ >> r.own_ns_tid_ >> r.clone_flags_;
       break;
     case TraceTaskEvent::EXEC:
       tasks >> r.file_name_ >> r.cmd_line_ >> r.fds_to_close_;
