+++ src/TraceTaskEvent.h
@@ -30,10 +30,11 @@ public:
 
   TraceTaskEvent(Type type = NONE, pid_t tid = 0) : type_(type), tid_(tid) {}
 
-  static TraceTaskEvent for_clone(pid_t tid, pid_t parent_tid,
+  static TraceTaskEvent for_clone(pid_t tid, pid_t parent_tid, pid_t own_ns_tid,
                                   uint64_t clone_flags) {
     TraceTaskEvent result(CLONE, tid);
     result.parent_tid_ = parent_tid;
+    result.own_ns_tid_ = own_ns_tid;
     result.clone_flags_ = clone_flags;
     return result;
   }
@@ -55,6 +56,10 @@ public:
   pid_t parent_tid() const {
     assert(type() == CLONE);
     return parent_tid_;
+  }
+  pid_t own_ns_tid() const {
+    assert(type() == CLONE);
+    return own_ns_tid_;
   }
   uint64_t clone_flags() const {
     assert(type() == CLONE);
@@ -89,6 +94,7 @@ private:
   Type type_;
   pid_t tid_;
   pid_t parent_tid_;                  // CLONE only
+  pid_t own_ns_tid_;                  // CLONE only
   uint64_t clone_flags_;              // CLONE only
   std::string file_name_;             // EXEC only
   std::vector<std::string> cmd_line_; // EXEC only
