+++ src/test/rr_ps_ns.run
@@ -1,0 +1,15 @@
+source `dirname $0`/util.sh
+compare_test EXIT-SUCCESS
+
+if !(do_ps | grep -q $'\t77'); then
+    # If 77 shows up, the test was skipped. Since we're here, we should expect
+    # the parent exiting with 79, and the child (PID 1 in its namespace) exiting
+    # with 78
+    if !(do_ps | grep -q $'\t79'); then
+      failed "Can't find return value 79"
+    fi
+    
+    if !(do_ps | grep $'\t78' | cut -f 1 | grep -q $'(1)'); then
+      failed "Can't find child namespace init"
+    fi    
+fi
